requirements.txt
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.2
django-environ==0.10.0
drf-yasg==1.21.6
coreapi==2.3.3
coverage==7.2.7
psycopg2==2.9.6
